- сохранение в БД uri-картинки

+++ app/src/main/java/ru/altarix/thegreatestnotes/model/DBHelper.java
@@ -15,9 +15,10 @@ public class DBHelper extends SQLiteOpenHelper {
     public static final String COLUMN_TITLE = "title";
     public static final String COLUMN_TEXT = "text";
     public static final String COLUMN_DATETIME = "datetime";
+    public static final String COLUMN_IMAGE_URI = "image_uri";
 
     private static final String DATABASE_NAME = "notes.db";
-    private static final int DATABASE_VERSION = 2;
+    private static final int DATABASE_VERSION = 3;
 
     // Database creation sql statement
     private static final String DATABASE_CREATE = "create table " + TABLE_NOTES
@@ -25,8 +26,13 @@ public class DBHelper extends SQLiteOpenHelper {
             + COLUMN_ID + " integer primary key autoincrement, "
             + COLUMN_TITLE + " text,"
             + COLUMN_TEXT + " text not null,"
-            + COLUMN_DATETIME + " int"
+            + COLUMN_DATETIME + " int,"
+            + COLUMN_IMAGE_URI + " text"
             + ");";
+
+    private static final String DATABASE_ALTER_FROM_2_TO_3 = "alter table " + TABLE_NOTES
+            + " add column "
+            + COLUMN_IMAGE_URI + " text";
 
     public DBHelper(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
@@ -39,10 +45,15 @@ public class DBHelper extends SQLiteOpenHelper {
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        Log.w(DBHelper.class.getName(),
-                "Upgrading database from version " + oldVersion + " to "
-                        + newVersion + ", which will destroy all old data");
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_NOTES);
-        onCreate(db);
+
+        if (oldVersion == 2 && newVersion == 3) {
+            db.execSQL(DATABASE_ALTER_FROM_2_TO_3);
+        }else {
+            Log.w(DBHelper.class.getName(),
+                    "Upgrading database from version " + oldVersion + " to "
+                            + newVersion + ", which will destroy all old data");
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_NOTES);
+            onCreate(db);
+        }
     }
 }
